feat(CLI): app cli implemented, file organizing, param type fixes

+++ app/models/Bill.py
@@ -3,6 +3,6 @@ from datetime import datetime
 
 class Bill:
     """Object that contains data about a bill."""
-    def __init__(self, amount: int, period: datetime) -> None:
+    def __init__(self, amount: float, period: datetime) -> None:
         self.amount = amount
         self.period = period
